requirements.txt
requests
openai
pandas
numpy
pillow
google-cloud-aiplatform
genai
matplotlib
google-cloud-storage